Skeleton for endpoints
