--- triAll/src/main/java/models/GameBoard.java
@@ -1,19 +1,0 @@
-package models;
-
-public class GameBoard {
-
-  private Player p1;
-
-  private Player p2;
-
-  private boolean gameStarted;
-
-  private int turn;
-
-  private char[][] boardState;
-
-  private int winner;
-
-  private boolean isDraw;
-
-}
