--- triAll/src/main/java/models/Move.java
@@ -1,11 +1,0 @@
-package models;
-
-public class Move {
-
-  private Player player;
-
-  private int moveX;
-
-  private int moveY;
-
-}
