--- triAll/src/main/java/models/Player.java
@@ -1,9 +1,0 @@
-package models;
-
-public class Player {
-
-  private char type;
-
-  private int id;
-
-}
