+++ triAll/src/main/java/models/Trial.java
@@ -1,0 +1,6 @@
+package models;
+
+public class Trial {
+	
+	private int id;
+}
