+++ triAll/src/main/java/models/User.java
@@ -1,0 +1,9 @@
+package models;
+
+import java.util.LinkedList;
+
+public class User {
+	private String username;
+	private boolean isResearcher;
+	private LinkedList<Trial> trials;
+}
